refactor(useBreadcrumbPath): Move into the breadcrumb feature
