+++ src/modules/views/Folder/FolderViewBreadcrumb.jsx
@@ -1,8 +1,8 @@
 import PropTypes from 'prop-types'
 import React, { useCallback } from 'react'
 
-import { useBreadcrumbPath } from './hooks/useBreadcrumbPath.jsx'
 import { MobileAwareBreadcrumb as Breadcrumb } from 'modules/breadcrumb/components/MobileAwareBreadcrumb'
+import { useBreadcrumbPath } from 'modules/breadcrumb/hooks/useBreadcrumbPath.jsx'
 
 const FolderViewBreadcrumb = ({
   currentFolderId,
