+++ src/modules/views/Folder/FolderViewBreadcrumb.spec.jsx
@@ -2,13 +2,13 @@ import { render } from '@testing-library/react'
 import React from 'react'
 
 import FolderViewBreadcrumb from './FolderViewBreadcrumb'
-import { useBreadcrumbPath } from './hooks/useBreadcrumbPath'
+import { useBreadcrumbPath } from 'modules/breadcrumb/hooks/useBreadcrumbPath'
 import {
   dummyBreadcrumbPath,
   dummyRootBreadcrumbPath
 } from 'test/dummies/dummyBreadcrumbPath'
 
-jest.mock('./hooks/useBreadcrumbPath')
+jest.mock('modules/breadcrumb/hooks/useBreadcrumbPath')
 jest.mock('modules/breadcrumb/components/MobileAwareBreadcrumb', () => ({
   // eslint-disable-next-line react/display-name
   MobileAwareBreadcrumb: ({ path, opening }) => (
